[FEATURE] 시도 횟수만큼 진행상황 출력
