+++ README.md
@@ -32,7 +32,7 @@
 - [ ] 실행 결과 출력
     - [x] 전진 횟수를 "-"로 출력
     - [x] 자동차 개수만큼 출력
-    - [ ] 시도 횟수만큼 출력
+    - [x] 시도 횟수만큼 출력
 - [ ] 최종 우승자 출력
 
 ##프로그래밍 요구사항
