+++ src/test/java/racingcar/domain/CarTest.java
@@ -25,19 +25,4 @@ public class CarTest {
         Car car = new Car("pobi");
         assertThat(car.movePosition(3)).isEqualTo(0);
     }
-
-    @Test
-    void 여러대의_차_진행상황_출력() {
-        List<Car> carList = new ArrayList<Car>();
-        carList.add(new Car("pobi"));
-        carList.add(new Car("elly"));
-        carList.add(new Car("rutgo"));
-        Cars cars = new Cars(carList);
-        cars.getCar(0).movePosition(5);
-        cars.getCar(0).movePosition(5);
-        cars.getCar(0).movePosition(5);
-        cars.getCar(1).movePosition(5);
-        cars.getCar(2).movePosition(5);
-        assertThat(cars.printCurrentResult()).isEqualTo("pobi : ---\nelly : -\nrutgo : -\n");
-    }
 }
