+++ src/test/java/racingcar/domain/GameTest.java
@@ -1,5 +1,11 @@
 package racingcar.domain;
 
-public class GameTest {
+import org.junit.jupiter.api.Test;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class GameTest {
 }
